deno to node (#60)
This changes the language runtime from deno to node. The reason for this change is because the target of this library is the browser and node has the best compatibility with it. Further we were running into a bunch of issues with the deno tooling for creating npm packages (e.g. dnt).

deno is also pretty good at loading npm packages into its runtime so starting in deno doesn't help us that much.

Ultimately there's just a few too many layers in between us and the npm release and at this point deno is not really serving us.

---------

Co-authored-by: Vlad <vladmarginean@softwiz.ro>